storybook: global language selector
