--- web/.storybook/preview.js
@@ -1,24 +1,0 @@
-import { withThemeByClassName } from '@storybook/addon-themes';
-
-import '../src/index.css';
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-};
-
-export const decorators = [
-  withThemeByClassName({
-    themes: {
-      light: '',
-      dark: 'dark',
-    },
-    defaultTheme: 'light',
-  }),
-];
-export const tags = ['autodocs'];
